20. Made changes to the models, adjusted the migrations, made changes to the filtering, added policies, and implemented checks for data manipulation permissions.

+++ app/Services/V1/CustomerService.php
@@ -8,7 +8,9 @@ use App\Dto\ApiListingDto;
 use App\Dto\Customer\CustomerListDto;
 use App\Dto\Customer\CustomerStoreDto;
 use App\Dto\Customer\CustomerUpdateDto;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\Auth;
 
 final class CustomerService implements CustomerServiceContract
 {
@@ -24,7 +26,7 @@ final class CustomerService implements CustomerServiceContract
 
     public function show(int $id): Model
     {
-        return $this->customerRepository->show($id);
+        return $this->customerRepository->find($id);
     }
 
     public function store(CustomerStoreDto $dto): Model
@@ -32,13 +34,33 @@ final class CustomerService implements CustomerServiceContract
         return $this->customerRepository->store($dto);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function update(CustomerUpdateDto $dto): Model
     {
+        $user  = Auth::user();
+        $model = $this->customerRepository->find($dto->id);
+
+        if (!$user->can('update', $model)) {
+            throw new AuthorizationException('No permissions for this action');
+        }
+
         return $this->customerRepository->update($dto);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function destroy(int $id): bool
     {
+        $user  = Auth::user();
+        $model = $this->customerRepository->find($id);
+
+        if (!$user->can('delete', $model)) {
+            throw new AuthorizationException('No permissions for this action');
+        }
+
         return $this->customerRepository->destroy($id);
     }
 }
